string: Merge strcat KUnit tests into string_kunit.c
mainline inclusion
from mainline-6.9-rc5
commit bd678f7d9b72ab8b6978dac92b841e46f4b935a3
category: feature
bugzilla: https://github.com/RVCK-Project/rvck/issues/280

--------------------------------

Move the strcat() tests into string_kunit.c. Remove the separate
Kconfig and Makefile rule.

Tested-by: Ivan Orlov <ivan.orlov0322@gmail.com>
Link: https://lore.kernel.org/r/20240419140155.3028912-4-keescook@chromium.org
